Refactor: Changing the directory structure
