+++ cmd/ft_activity_api/main.go
@@ -7,6 +7,8 @@ import (
 	"os"
 	"github.com/gin-gonic/gin"
 	"github.com/gin-contrib/cors"
+	"github.com/42association/42ActivityAPI/internal/handlers"
+	"github.com/42association/42ActivityAPI/internal/database"
 )
 
 type Config struct {
