+++ cmd/ft_activity_api/main_test.go
@@ -10,6 +10,8 @@ import (
 	"net/http/httptest"
 	"net/http"
 	"os"
+	"github.com/42association/42ActivityAPI/internal/handlers"
+	"github.com/42association/42ActivityAPI/internal/database"
 )
 
 func setupTestDB() *gorm.DB {
